+++ src/Solenoid.h
@@ -4,7 +4,7 @@
 class Solenoid
 {
 private:
-    int activationNum;  //The number that gets fed to the shift registers to activate this specific solenoid's relay (to pour)
+    short int activationNum;  //The number that gets fed to the shift registers to activate this specific solenoid's relay (to pour)
     short int xPos;     //The number representing the fixed location of the solenoid. "How many revolutions away from the stepper motor am I"
     
 
@@ -13,7 +13,7 @@ public:
     Solenoid();
 
     //Parameterized constructor
-    Solenoid(int arg_activationNum, short int arg_xPos);
+    Solenoid(short int arg_activationNum, short int arg_xPos);
 
     // Opens solenoid (pours) for X milliseconds.
     void Pour(int milliseconds);
@@ -24,6 +24,6 @@ public:
 
     // Setters
     void setxPos(short int arg_xPos);
-    void setactivationNum(int arg_activationNum);
+    void setactivationNum(short int arg_activationNum);
 };
 #endif
